app-server: resolve Unix socket path at parse time
